My department employees data related code moved into reducers/people/*. reducers/user/* cleaned up.
