+++ mobile/src/reducers/app.reducer.ts
@@ -15,6 +15,7 @@ import { UserInfoState, userInfoReducer } from './user/user-info.reducer';
 import { userEpics } from './user/user.reducer';
 import { FeedsState, feedsReducer, feedsEpics } from './feeds/feeds.reducer';
 import { CalendarState, calendarReducer } from './calendar/calendar.reducer';
+import { PeopleState, peopleReducer, peopleEpics } from './people/people.reducer';
 
 export interface AppState {
     helpdesk: HelpdeskState;
@@ -23,9 +24,10 @@ export interface AppState {
     userInfo: UserInfoState;
     feeds: FeedsState;
     calendar: CalendarState;
+    people: PeopleState;
 }
 
-const rootEpic = combineEpics(helpdeskEpics as any, organizationEpics as any, errorsEpics as any, userEpics as any, feedsEpics as any);
+const rootEpic = combineEpics(helpdeskEpics as any, organizationEpics as any, errorsEpics as any, userEpics as any, feedsEpics as any, peopleEpics as any);
 
 const reducers = combineReducers<AppState>({
     helpdesk: helpdeskReducer,
@@ -33,7 +35,8 @@ const reducers = combineReducers<AppState>({
     nav: navigationReducer,
     userInfo: userInfoReducer,
     feeds: feedsReducer,
-    calendar: calendarReducer
+    calendar: calendarReducer,
+    people: peopleReducer
 });
 
 export const storeFactory = () => {
