+++ mobile/src/reducers/user/user.action.ts
@@ -22,11 +22,4 @@ export interface LoadUserEmployeeFinished {
 
 export const loadUserEmployeeFinished = (employee: Employee): LoadUserEmployeeFinished => ({ type: 'LOAD-USER-EMPLOYEE-FINISHED', employee });
 
-export interface LoadUserDepartmentEmployessFinished {
-    type: 'LOAD-USER-DEPARTMENT-EMPOYEES-FINISHED';
-    employee: Employee;
-}
-
-export const loadUserDepartmentEmployessFinished = (employee: Employee): LoadUserDepartmentEmployessFinished => ({ type: 'LOAD-USER-DEPARTMENT-EMPOYEES-FINISHED', employee });
-
-export type UserActions = LoadUser | LoadUserFinished | LoadUserEmployeeFinished | LoadUserDepartmentEmployessFinished;
+export type UserActions = LoadUser | LoadUserFinished | LoadUserEmployeeFinished;
